backend/app/schemas/catalog.py
from pydantic import BaseModel, ConfigDict


class PackagingOptionResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    name: str
    price: int


class ExtraServiceResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    name: str
    price: int
