backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
email-validator==2.3.0
